backend/pyproject.toml
[project]
name = "digital-fte-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "asyncpg>=0.30",
    "pydantic-settings>=2.5",
    "openai-agents>=0.0.1",
    "structlog>=24.0",
    "python-dotenv>=1.0",
    "aiokafka>=0.13.0",
    "openai>=2.30.0",
    "pgvector>=0.4.2",
    "numpy>=2.4.3",
    "google-auth-oauthlib>=1.3.0",
    "google-auth-httplib2>=0.3.0",
    "google-api-python-client>=2.193.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
    "mypy>=1.13",
    "locust>=2.43.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
